bas2d-agent/test-requirements.txt
attrs==21.4.0
avikus-amq==0.4.2
avikus-event-loop==0.3.1
Deprecated==1.2.13
iniconfig==1.1.1
#numpy==1.19.4
#opencv-python==4.5.5.64
packaging==21.3
Pillow>=8.4.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
python-dotenv==0.20.0
pyzmq==22.3.0
redis==4.1.4
simplejpeg==1.6.4
tomli==1.2.3
wrapt==1.14.0
zmq==0.0.0
